requirements.txt
# Environment variable loading
python-dotenv>=1.0.0

# Snowflake dependencies
snowflake>=1.11
snowflake-snowpark-python>=1.11.0
snowflake-connector-python[secure-local-storage]>=3.6.0

# Google ADK (includes google-genai as dependency)
google-adk[vertexai]>=1.27.2

# LiteLLM — enables Anthropic, OpenAI, and 100+ other model providers
litellm>=1.40.0
google-cloud-aiplatform>=1.110.0
openai>=1.0.0
anthropic>=0.34.0
google-generativeai>=0.8.0
requests>=2.31.0

# Google Model Armor (prompt sanitization)
google-cloud-modelarmor>=0.4.0

# Data validation
pydantic>=2.5.0

# Version control
GitPython>=3.1.40

# Cron expression parsing
croniter>=2.0.0

# Additional utility packages (if needed)
python-dateutil>=2.8.2

# OpenTelemetry — enabled via OTEL_ENABLED=true in .env
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.25.0
opentelemetry-instrumentation-logging>=0.46b0

# Rich terminal UI
rich>=13.0.0

# Interactive terminal input
prompt_toolkit>=3.0.0

# for vertexai
google-cloud-aiplatform>=1.135.0

# For backend
fastapi
uvicorn
pydantic
pydantic_settings
pymongo
authlib
itsdangerous
httpx
pytest
